+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfb68tc18dyv3hsncb31kyjn"
+        "conv_01kfbbz05kaqhs7atv1wtqmnec"
     )
     r"""Unique conversation identifier with `conv_` prefix."""