+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFB68V3YTTTG0APTCMXX2S9F"
+        "01KFBBZ0N9R80JRAG45H3545SZ"
     )
     r"""The unique identifier of the data source"""