+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1723,7 +1723,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KFB68T7T10AZ2S35PBRDHRC3"
+    id: Optional[str] = "01KFBBZ0059P3Y79KT429AB4Z7"
 
     description: Optional[str] = None