+++ packages/orq-rc/RELEASES.md
@@ -4789,3 +4789,13 @@ Based on:
 - [python v4.2.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.41 - packages/orq-rc
+
+## 2026-01-19 14:56:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.42 - packages/orq-rc